pyproject.toml
[project]
name = "kawaiige-mu"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.2.2",
    "supabase>=2.28.0",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]

[dependency-groups]
dev = [
    "ruff>=0.15.5",
]
